Specify node version requirements in package.json
The latest release to 3.0.5 introduced es6 syntax which makes using base-x difficult as a dependency. This has been the case with me and I'm sure with other developers that have not yet made the switch over to es6. 

Introducing constraint so that npm is aware that 3.0.5 only works with node versions greater than 6.