feature: remove node Buffers and return Uint8Arrays
Instead of returning node Buffers, this PR changes the API to return
Uint8Arrays.

The node Buffer polyfill is large and it's not included by default with
webpack any more.

The parts of the Buffer API that most modules use are covered by the
browser-native [DataView](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView)
class, and of course Buffer is a subclass of Uint8Array so there's isn't
a very compelling reason to use Buffers in the future.